requirements.txt
django
python-dotenv
django-cors-headers